Move transmission_adaptor to mhs_common
