+++ mhs/common/mhs_common/workflow/__init__.py
@@ -2,9 +2,9 @@
 from typing import Dict
 
 from comms import queue_adaptor
-from comms import transmission_adaptor
 
 from mhs_common.state import persistence_adaptor
+from mhs_common.transmission import transmission_adaptor
 from mhs_common.workflow.asynchronous_express import AsynchronousExpressWorkflow
 from mhs_common.workflow.asynchronous_reliable import AsynchronousReliableWorkflow
 from mhs_common.workflow.common import CommonWorkflow
